reactor: reservation 관련 코드 reservation 패키지로 이동
 FIX
    - reservation 관련 코드 reservation 패키지로 이동
