+++ src/main/java/com/example/livealone/broadcast/controller/BroadcastController.java
@@ -4,17 +4,12 @@ import com.example.livealone.broadcast.dto.BroadcastRequestDto;
 import com.example.livealone.broadcast.dto.BroadcastResponseDto;
 import com.example.livealone.broadcast.dto.BroadcastTitleResponseDto;
 import com.example.livealone.broadcast.dto.CreateBroadcastResponseDto;
-import com.example.livealone.broadcast.dto.ReservationStateResponseDto;
-import com.example.livealone.broadcast.dto.ReservationRequestDto;
-import com.example.livealone.broadcast.dto.ReservationResponseDto;
 import com.example.livealone.broadcast.dto.UserBroadcastResponseDto;
 import com.example.livealone.broadcast.service.BroadcastService;
 import com.example.livealone.global.dto.CommonResponseDto;
-import com.example.livealone.global.mail.MailService;
 import com.example.livealone.global.security.UserDetailsImpl;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import jakarta.validation.Valid;
-import java.time.LocalDate;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -33,7 +28,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class BroadcastController {
 
   private final BroadcastService broadcastService;
-  private final MailService mailService;
 
   @PostMapping("/broadcast")
   public ResponseEntity<CommonResponseDto<CreateBroadcastResponseDto>> createBroadcast(
@@ -92,35 +86,6 @@ public class BroadcastController {
             null)
     );
 
-  }
-
-  @PostMapping("/broadcast/reservation")
-  public ResponseEntity<CommonResponseDto<Void>> createReservation(
-      @AuthenticationPrincipal UserDetailsImpl userPrincipal,
-      @RequestBody ReservationRequestDto requestDto) {
-
-    ReservationResponseDto responseDto = broadcastService.createReservation(requestDto, userPrincipal.getUser());
-
-    mailService.sendEmail(responseDto.getEmail(),
-        "[LiveAlone]방송 예약 완료 확인 메일입니다.",
-        responseDto.getAirTime() + "타임 방송 예약\n streamKey: " + responseDto.getCode());
-
-    return ResponseEntity.status(HttpStatus.OK).body(
-        new CommonResponseDto<>(
-            HttpStatus.OK.value(),
-            "예약을 성공하였습니다. 이메일은 순차적으로 발송됩니다.",
-            null)
-    );
-  }
-
-  @GetMapping("/broadcast/reservations")
-  public ResponseEntity<CommonResponseDto<List<ReservationStateResponseDto>>> getReservations(@RequestParam LocalDate date) {
-    return ResponseEntity.status(HttpStatus.OK).body(
-        new CommonResponseDto<>(
-            HttpStatus.OK.value(),
-            "예약 목록 조회를 성공하였습니다.",
-            broadcastService.getReservations(date))
-    );
   }
 
   @GetMapping("/broadcast/{broadcastId}")
