--- src/main/java/com/example/livealone/broadcast/dto/ReservationRequestDto.java
@@ -1,11 +1,0 @@
-package com.example.livealone.broadcast.dto;
-
-import java.time.LocalDateTime;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class ReservationRequestDto {
-  private LocalDateTime airtime;
-}
