+++ src/main/java/com/example/livealone/broadcast/repository/BroadcastRepository.java
@@ -2,7 +2,7 @@ package com.example.livealone.broadcast.repository;
 
 import com.example.livealone.broadcast.entity.Broadcast;
 import com.example.livealone.broadcast.entity.BroadcastStatus;
-import com.example.livealone.broadcast.entity.Reservations;
+import com.example.livealone.reservation.entity.Reservations;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
