+++ src/main/java/com/example/livealone/reservation/dto/ReservationRequestDto.java
@@ -1,0 +1,11 @@
+package com.example.livealone.reservation.dto;
+
+import java.time.LocalDateTime;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class ReservationRequestDto {
+  private LocalDateTime airtime;
+}
