+++ src/main/java/com/example/livealone/reservation/dto/ReservationResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.livealone.broadcast.dto;
+package com.example.livealone.reservation.dto;
 
 import java.time.LocalDateTime;
 import lombok.Builder;
