+++ src/main/java/com/example/livealone/reservation/repository/ReservationRepository.java
@@ -1,0 +1,14 @@
+package com.example.livealone.reservation.repository;
+
+import com.example.livealone.reservation.entity.Reservations;
+import com.example.livealone.user.entity.User;
+import java.time.LocalDateTime;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ReservationRepository extends JpaRepository<Reservations, Long>, ReservationRepositoryQuery {
+  Optional<Reservations> findByAirTime(LocalDateTime airtime);
+
+  Optional<Reservations> findByAirTimeBetweenAndStreamer(LocalDateTime start, LocalDateTime end, User user);
+
+}
